requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
aiohttp>=3.9.0
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
Pillow>=10.2.0
APScheduler>=3.10.4
httpx>=0.27.0
python-dotenv>=1.0.0
piexif>=1.1.3
